Restructure gnw interface code
