--- pydiffexp/__init__.py
@@ -1,5 +1,0 @@
-from .legacy import *
-from .diffexp import *
-from .diffexp import DEAnalysis
-from .plot import *
-from .GNWwrapper import *
