+++ pydiffexp/gnw/__init__.py
@@ -1,0 +1,2 @@
+from .GnwWrapper import GnwNetwork
+from .sim_explorer import *
